// src/App/components/NavMenu/NavMenu.d.ts
import React from "react";

// What gets passed into the component from the parent as attributes
export declare interface INavMenuInputProps {

}

export type NavMenuProps = INavMenuInputProps;

export declare interface ILink {
    url: string;
    icon: JSX.Element;
    text: string;
}